# static/specs/openapi/catalog/catalog_agent.yaml
openapi: 3.0.3
info:
  title: Rainbow Catalog Agent API - Entity Management
  description: OpenAPI specification for the Rainbow Catalog Agent Entity Management endpoints.
  version: 1.0.0
servers:
  - url: /v1/catalog-agent
    description: Base path for the catalog agent API
...
security:
  - BearerAuth: []
paths:
...
  /catalogs/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags:
        - Catalogs
      summary: Get catalog by ID
      operationId: getCatalogById
      responses:
        "200":
          description: Catalog found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Catalogs
      summary: Update catalog by ID
      operationId: updateCatalogById
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditCatalogDto"
            example:
              dctTitle: "Updated Catalog Title"
      responses:
        "202":
          description: Catalog updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Catalogs
      summary: Delete catalog by ID
      operationId: deleteCatalogById
      responses:
        "202":
          description: Catalog deleted (Accepted)
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --- Data Services ---
...
  /data-services/catalog/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags:
        - Data Services
      summary: Get data services by catalog ID
      operationId: getDataServicesByCatalogId
      responses:
        "200":
          description: List of data services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataServiceDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /datasets/catalog/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags:
        - Datasets
      summary: Get datasets by catalog ID
      operationId: getDatasetsByCatalogId
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DatasetDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /odrl-policies:
    get:
      tags:
        - ODRL Policies
      summary: Get all ODRL policies
      operationId: getAllOdrlPolicies
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: List of ODRL policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OdrlPolicyDto"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - ODRL Policies
      summary: Create ODRL policy
      operationId: createOdrlPolicy
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewOdrlPolicyDto"
            example:
              entityId: "urn:uuid:my-dataset-id"
              entityType: "Dataset"
              odrlOffer:
                profile: ["http://www.w3.org/ns/odrl/2/"]
                permission:
                  - action: "http://www.w3.org/ns/odrl/2/use"
      responses:
        "200":
          description: ODRL policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OdrlPolicyDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /odrl-policies/batch:
    post:
      tags:
        - ODRL Policies
      summary: Get batch ODRL policies
      operationId: getBatchOdrlPolicies
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRequests"
      responses:
        "200":
          description: List of ODRL policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OdrlPolicyDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /odrl-policies/entity/{entity_id}:
    parameters:
      - name: entity_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - ODRL Policies
      summary: Get ODRL policies by entity ID
      operationId: getOdrlPoliciesByEntity
      responses:
        "200":
          description: List of ODRL policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OdrlPolicyDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - ODRL Policies
      summary: Delete ODRL policies by entity ID
      operationId: deleteOdrlPoliciesByEntity
      responses:
        "202":
          description: Policies deleted (Accepted)
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /odrl-policies/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags:
        - ODRL Policies
      summary: Get ODRL policy by ID
      operationId: getOdrlPolicyById
      responses:
        "200":
          description: ODRL policy found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OdrlPolicyDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - ODRL Policies
      summary: Delete ODRL policy by ID
      operationId: deleteOdrlPolicyById
      responses:
        "202":
          description: Policy deleted (Accepted)
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --- Policy Templates ---
...
  /peer-catalogs/{peer_id}:
    parameters:
      - name: peer_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Peer Catalogs
      summary: Get remote peer catalog
      operationId: getPeerCatalog
      responses:
        "200":
          description: Peer catalog found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        format: int64
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        format: int64
...
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
  schemas:
...
    ErrorInfo:
      type: object
      required:
        - message
        - error_code
        - cause
      properties:
        message:
          type: string
        error_code:
          type: integer
        details:
          type: string
        cause:
          type: string
      example:
        message: "Resource not found"
        error_code: 3200
        status_code: 404
        cause: "Catalog with ID urn:uuid:123 not found"

    # --- Common Wrappers ---
    BatchRequests:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
            format: urn
...
    CatalogDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        foafHomePage:
          type: string
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctIdentifier:
          type: string
        dctIssued:
          type: string
          format: date-time
        dctModified:
          type: string
          format: date-time
        dctTitle:
          type: string
        dspaceParticipantId:
          type: string
        dspaceMainCatalog:
          type: boolean
      example:
        {
          "dctConformsTo": null,
          "dctCreator": null,
          "dctIdentifier": "urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635",
          "dctIssued": "2026-01-09T10:49:53.822684Z",
          "dctModified": null,
          "dctTitle": null,
          "dspaceMainCatalog": true,
          "dspaceParticipantId": "did:jwk:eyJrdHkiOiJSU0Ei...",
          "foafHomePage": null,
          "id": "urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635"
        }
...
    EditCatalogDto:
      type: object
      properties:
        foafHomePage:
          type: string
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctTitle:
          type: string
      example:
        {
          "dctConformsTo": "string",
          "dctCreator": "string",
          "dctTitle": "string",
          "foafHomePage": "string",
        }

    # --- Data Services ---
    DataServiceDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        dcatEndpointDescription:
          type: string
        dcatEndpointUrl:
          type: string
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctIdentifier:
          type: string
        dctIssued:
          type: string
          format: date-time
        dctModified:
          type: string
          format: date-time
        dctTitle:
          type: string
        dctDescription:
          type: string
        catalogId:
          type: string
          format: urn
        dspaceMainDataService:
          type: boolean
      example:
        {
          "catalogId": "urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635",
          "dcatEndpointDescription": null,
          "dcatEndpointUrl": "http://127.0.0.1:1200/dsp/current",
          "dctConformsTo": null,
          "dctCreator": null,
          "dctDescription": null,
          "dctIdentifier": "urn:data-service:37840ca3-79f6-4542-8394-0c57dd93650f",
          "dctIssued": "2026-01-09T10:49:53.895893Z",
          "dctModified": null,
          "dctTitle": null,
          "dspaceMainDataService": true,
          "id": "urn:data-service:37840ca3-79f6-4542-8394-0c57dd93650f"
        }
...
    DatasetDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctIdentifier:
          type: string
        dctIssued:
          type: string
          format: date-time
        dctModified:
          type: string
          format: date-time
        dctTitle:
          type: string
        dctDescription:
          type: string
        catalogId:
          type: string
          format: urn
      example:
        {
          "catalogId": "urn:catalog:4958f37d-17cf-44e7-ae18-1f26b6fe15ff",
          "dctConformsTo": null,
          "dctCreator": null,
          "dctDescription": null,
          "dctIdentifier": "urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec",
          "dctIssued": "2026-01-10T14:23:29.091150Z",
          "dctModified": null,
          "dctTitle": "superfoaf",
          "id": "urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec"
        }
...
    OdrlPolicyDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        odrlOffer:
          $ref: "#/components/schemas/OdrlPolicyInfo"
        entity:
          type: string
          format: urn
        entityType:
          type: string
          enum: [Distribution, DataService, Catalog, Dataset]
        createdAt:
          type: string
          format: date-time
        sourceTemplateId:
          type: string
        sourceTemplateVersion:
          type: string
        instantiationParameters:
          type: object
      example:
        {
          "createdAt": "2026-01-10T14:40:30.978559Z",
          "entity": "urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635",
          "entityType": "Catalog",
          "id": "urn:odrl-policy:abad083e-78a8-4f90-aaca-7c5d6af6c251",
          "instantiationParameters": {
            "$bla": "hola"
          },
          "odrlOffer": {
            "permission": [
              {
                "action": "use",
                "constraint": [
                  {
                    "leftOperand": "bla",
                    "operator": "eq",
                    "rightOperand": "hola"
                  }
                ]
              }
            ]
          },
          "sourceTemplateId": "policy-1",
          "sourceTemplateVersion": "1.0"
        }
    NewOdrlPolicyDto:
      type: object
      required:
        - odrlOffer
        - entityId
        - entityType
      properties:
        id:
          type: string
          format: urn
        odrlOffer:
          $ref: "#/components/schemas/OdrlPolicyInfo"
        entityId:
          type: string
          format: urn
        entityType:
          type: string
          enum: [Distribution, DataService, Catalog, Dataset]
        sourceTemplateId:
          type: string
        sourceTemplateVersion:
          type: string
        instantiationParameters:
          type: object
      example:
        {
          "entity": "urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635",
          "entityType": "Catalog",
          "odrlOffer": {
            "permission": [
              {
                "action": "use",
                "constraint": [
                  {
                    "leftOperand": "bla",
                    "operator": "eq",
                    "rightOperand": "hola"
                  }
                ]
              }
            ]
          }
        }
    OdrlPolicyInfo:
      type: object
      properties:
        profile:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        permission:
          type: array
          items:
            $ref: "#/components/schemas/OdrlPermission"
        obligation:
          type: array
          items:
            $ref: "#/components/schemas/OdrlObligation"
        prohibition:
          type: array
          items:
            $ref: "#/components/schemas/OdrlObligation"
    OdrlPermission:
      type: object
      required:
        - action
      properties:
        action:
          type: string
        constraint:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"
        duty:
          $ref: "#/components/schemas/OdrlDuty"
    OdrlDuty:
      type: object
      required:
        - action
      properties:
        action:
          type: string
        constraint:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"
    OdrlObligation:
      $ref: "#/components/schemas/OdrlDuty"
    OdrlConstraint:
      oneOf:
        - $ref: "#/components/schemas/OdrlAtomicConstraint"
        - $ref: "#/components/schemas/OdrlLogicalConstraint"
    OdrlAtomicConstraint:
      type: object
      required:
        - rightOperand
        - leftOperand
        - operator
      properties:
        leftOperand:
          type: string
        operator:
          type: string
          enum:
            [
              eq,
              gt,
              gteq,
              lteq,
              hasPart,
              isA,
              isAllOf,
              isAnyOf,
              isNoneOf,
              isPartOf,
              lt,
              termLteq,
              neq,
            ]
        rightOperand:
          oneOf:
            - type: string
            - type: object
            - type: array
              items: {}
    OdrlLogicalConstraint:
      type: object
      properties:
        and:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"
        andSequence:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"
        or:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"
        xone:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"

    # --- Policy Templates ---
...
    Catalog:
      type: object
      properties:
        "@context":
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        "@type":
          type: string
        "@id":
          type: string
          format: urn
        homepage:
          type: string
        theme:
          type: string
        keyword:
          type: string
        conformsTo:
          type: string
        creator:
          type: string
        identifier:
          type: string
        issued:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: array
          items:
            type: string
        participantId:
          type: string
        hasPolicy:
          type: array
          items:
            $ref: "#/components/schemas/OdrlOffer"
        extraFields:
          type: object
        catalog:
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/Catalog"
            - type: array
              items:
                $ref: "#/components/schemas/CatalogMinimized"
        dataset:
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/Dataset"
            - type: array
              items:
                $ref: "#/components/schemas/DatasetMinimized"
        service:
          oneOf:
            - $ref: "#/components/schemas/DataService"
            - $ref: "#/components/schemas/DataServiceMinimized"
            - type: array
              items:
                $ref: "#/components/schemas/DataService"
    CatalogMinimized:
      type: object
      properties:
        "@type":
          type: string
        "@id":
          type: string
          format: urn
        homepage:
          type: string
    Dataset:
      type: object
      properties:
        "@context":
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        "@type":
          type: string
        "@id":
          type: string
        hasPolicy:
          type: array
          items:
            $ref: "#/components/schemas/OdrlOffer"
        distribution:
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/Distribution"
            - type: array
              items:
                $ref: "#/components/schemas/DistributionMinimized"
        conformsTo:
          type: string
    DatasetMinimized:
      type: object
      properties:
        "@type":
          type: string
        "@id":
          type: string
    DataService:
      type: object
      properties:
        "@context":
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        "@type":
          type: string
        "@id":
          type: string
        endpointURL:
          type: string
        endpointDescription:
          type: string
    DataServiceMinimized:
      type: object
      properties:
        "@type":
          type: string
        "@id":
          type: string
    Distribution:
      type: object
      properties:
        "@type":
          type: string
        "@id":
          type: string
        accessService:
          type: array
          items:
            $ref: "#/components/schemas/DataService"
    DistributionMinimized:
      type: object
      properties:
        "@type":
          type: string
        "@id":
          type: string
    OdrlOffer:
      type: object
      properties:
        "@type":
          type: string
          enum: [Offer, Agreement]
        "@id":
          type: string
          format: urn
        target:
          type: string
          format: urn
        profile:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        permission:
          type: array
          items:
            $ref: "#/components/schemas/OdrlPermission"
        obligation:
          type: array
          items:
            $ref: "#/components/schemas/OdrlObligation"
        prohibition:
          type: array
          items:
            $ref: "#/components/schemas/OdrlObligation"